+++ frontend/src/components/molecules/ChartCards/StableCoinDistributionChart.vue
@@ -3,7 +3,7 @@
 		<DistributionChart
 			:is-loading="isLoading"
 			:stable-coins-data="delayedSupplyPercentage"
-			:label-clickable="true"
+			:label-clickable="false"
 		>
 			<template #title>Protocol Token Supply Distribution </template>
 		</DistributionChart>
@@ -32,7 +32,11 @@ const props = defineProps<Props>()
 const delayedSupplyPercentage = ref<StableCoin[]>([])
 
 watchEffect(() => {
-	const stablecoins = props.stableCoinsData?.stablecoins || []
+	const stablecoins = props.stableCoinsData?.liveStablecoins
+		? [...props.stableCoinsData.liveStablecoins].sort(
+				(a, b) => (b.totalSupply ?? 0) - (a.totalSupply ?? 0)
+		  )
+		: []
 
 	if (stablecoins.length === 0) {
 		delayedSupplyPercentage.value = []
@@ -44,12 +48,11 @@ watchEffect(() => {
 			(sum, coin) => sum + (coin.totalSupply || 0),
 			0
 		)
-
 		delayedSupplyPercentage.value = stablecoins.map(coin => ({
 			symbol: coin.symbol,
 			supplyPercentage:
 				totalSupplySum > 0
-					? ((coin.totalSupply! / totalSupplySum) * 100).toFixed(2)
+					? (((coin.totalSupply ?? 0) / totalSupplySum) * 100).toFixed(2)
 					: '0',
 		}))
 	}, 1000)