+++ backend/CHANGELOG.md
@@ -4,6 +4,28 @@ All notable changes to this project will be documented in this file.
 
 ## Unreleased
 
+## [2.0.13] - 2025-07-03
+
+### Changed
+
+Database schema version: 38
+
+- Baker APY query db migration to change the function to prevent overflow FLOAT8
+
+## [2.0.12] - 2025-07-03
+
+### Changed
+
+- New account_transaction_type enum values added `TokenUpdate`, Removed `TokenHolder` and `TokenGovernance`
+- TokenHolder and TokenGovernance events are now merged into `TokenUpdate` event.
+- The ccdscan-api now connects to the grpc node to fetch token list and details (This will be removed later).
+
+## [2.0.11] - 2025-07-01
+
+### Fixed
+
+- Fix for db connections issue where new connections are created during failed block processing, and old connections are not closed and cleaned up effectively
+
 ## [2.0.10] - 2025-06-24
 
 ### Added